Extracting the token generation and comparator.
Also renaming the ModelAuthentication#token_expired? method to
prevent overlaping method names.
